Use the new works results card on works search
